Push mplayer into a child process
mplayer is now executed in a separate child process (feeding
another child process for ffmpeg, as before).  This prevents the
program from blocking indefinitely when mplayer fills up the pipe
buffer and where ffmpeg crashes.

Under the new strategy, failure in either ffmpeg or mplayer will
not stop the top-level program from running to completion.

Also, this commit removes the dot-file approach to hiding the fifo
created for the encode.  I don't bother hiding it anymore.  I'm
not hiding the intermediate file, and want it to be easy to manually
clean up if the program fails catastrophically without doing so
itself.