Remove the dependency on Log::Log4perl. 'generate' number pattern.
Log::Log4perl is too much for a script like this, and it's not worth the hassle
as the only non-core dependency.  Replaced with an internal logger.

The --number-pattern argument can now accept the string 'generate', in which
case transcoded videos are written out with an incrementing suffix.

Remove duplicate 'my' for $number variable.

Not fatal or dangerous, just the source of an annoying warning.

Updated README